requirements.txt
numpy
pandas
plotly
streamlit
ipykernel
python-dotenv
supabase
matplotlib
openai
Pillow